+Candidate 60%

+++ src/RMI/RMIinterface.java
@@ -14,13 +14,13 @@ public interface RMIinterface extends Remote{
      * @term::long   currentTerm for candidate to update itself
      * @voteGranted::boolean true means candidate grant vote
      */ 
-    ArrayList RequestVote(long term, int candidateId , int lastLogIndex , int lastLogTerm );
+    ArrayList RequestVote(long term, int candidateId , int lastLogIndex , long lastLogTerm );
     /** 
      * return ArraryList:
      * @term::long    currentTerm for leader to update itself
      * @success:boolean true if follower contained entry matching prevLogIndex and prevLogterm
      */ 
-    ArrayList AppendEntries(long term,int leaderId,int prevLogIndex,int prevLogTerm,
+    ArrayList AppendEntries(long term,int leaderId,int prevLogIndex,long prevLogTerm,
             ArrayList<Entry> entries,int leaderCommit);
    
 }
